backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
psycopg2-binary>=2.9,<3
gunicorn>=21,<23
reportlab>=4.0,<5
python-ldap>=3.4; sys_platform != "win32"
